Makes a measurement metadata file

Make a measure_info.json file, or add measure entries to an existing one.

Usage

data_measure_info(path, ..., info = list(), references = list(),
  strict = FALSE, include_empty = TRUE, overwrite_entry = FALSE,
  overwrite = FALSE, verbose = TRUE, open_after = interactive())

Arguments

path

Path to the measure_info.json file, existing or to be created.

...

Lists containing individual measure items. See the Measure Entries section.

info

A list containing measurement information to be added.

references

A list containing citation entries. See the Reference Entries section.

strict

Logical; if TRUE, will only allow recognized entries and values.

include_empty

Logical; if FLASE, will omit entries that have not been provided.

overwrite_entry

Logical; if TRUE, will replace rather than add to an existing entry.

overwrite

Logical; if TRUE, will overwrite rather than add to an existing path.

verbose

Logical; if FALSE, will not display status messages.

open_after

Logical; if FALSE, will not open the measure file after writing/updating.

Value

An invisible list containing measurement metadata.

Measure Entries

Measure entries are named by the full variable name with any of these entries (if strict):

measure: Name of the measure.

full_name: Full name of the measure, which is also the name of the entry.

short_name: Shortest possible display name.

long_name: Longer display name.

category: Arbitrary category for the measure.

short_description: Shortest possible description.

long_description: Complete description.

statement: String with dynamic references to entity features (e.g., "measure value = {value}"). References can include:

value: Value of a currently displaying variable at a current time.

region_name: Alias of feature.name.

feature.<entry>: An entity feature, coming from entity_info.json or GeoJSON properties. All entities have at least name and id entries (e.g., "{feature.id}").

variables.<entry>: A variable feature such as name which is the same as full_name (e.g., "{variables.name}").

data.<variable>: The value of another variable at a current time (e.g., "{data.variable_a}").

type: Type of the measure's value. Recognized types are displayed in a special way:

year or integer show as entered (usually as whole numbers). Other numeric types are rounded to show a set number of digits.

percent shows as {value}%.

minutes shows as {value} minutes.

dollar shows as ${value}.

internet speed shows as {value} MB/s.

sources: A list or list of list containing source information, including any of these entries:

name: Name of the source (such as an organization name).

url: General URL of the source (such as an organization's website).

location: More specific description of the source (such as a the name of a particular data product).

location_url: More direct URL to the resource (such as a page listing data products).

date_accessed: Date of retrieval (arbitrary format).

citations: A vector of reference ids (the names of reference entries; e.g., c("ref1", "ref3")).

layer: A list specifying an output_map overlay:

source (required): A URL to a GeoJSON file, or a list with a url and time entry, where time conditions the display of the layer on the current selected time. Alternative to a list that specifies time, the URL can include a dynamic reference to time, if the time values correspond to a component of the URL (e.g., "https://example.com/{time}/points.geojson").

filter: A list or list of lists specifying how the elements of the layer should be filtered for this variable:

feature: Name of the layer's property to filter on.

operator: Operator to filter by (e.g., "=" or "!=").

value: Value to filter by.

Reference Entries

Reference entries can be included in a _references entry, and should have names corresponding to those included in any of the measures' citation entries. These can include any of these entries:

id: The reference id, same as the entry name.

author: A list or list of lists specifying one or more authors. These can include entries for given and family names.

year: Year of the publication.

title: Title of the publication.

journal: Journal in which the publication appears.

volume: Volume number of the journal.

page: Page number of the journal.

doi: Digital Object Identifier, from which a link is made (https://doi.org/{doi}).

version: Version number of software.

url: Link to the publication, alternative to a DOI.

Examples

path <- tempfile()

# make an initial file
data_measure_info(path, "measure name" = list(
  measure = "measure name",
  full_name = "prefix:measure name",
  short_description = "A measure.",
  statement = "This entity has {value} measure units."
), verbose = FALSE)

# add another measure to that
measure_info <- data_measure_info(path, "measure two" = list(
  measure = "measure two",
  full_name = "prefix:measure two",
  short_description = "Another measure.",
  statement = "This entity has {value} measure units."
), verbose = FALSE)
names(measure_info)
#> [1] "measure name" "measure two"