Add a number input to a website

Adds an direct number input element to a website.

input_number(label, id = label, ..., default = NULL, min = NULL,
  max = NULL, step = NULL, type = "number", class = NULL,
  dataview = NULL)

Arguments

label
Label of the input for the user.

id
Unique ID of the element to be created.

...
Other attributes to add to the input.

default
Default value of the input, the ID of an input to use as the default, or "min" or "max" to default to the current minimum or maximum value.

min
The smallest allowed value.

max
The largest allowed value.

step
Amount to increase or decrease the value by when changed with arrows.

type
Name of the input's type -- other number-based types like date might make sense.

class
Class names to add to the input's list.

dataview
ID of a input_dataview, to use as a source of variables.

Value

A character vector of the contents to be added.

Examples

if (FALSE) {
input_text("entered_text", "Enter Text:")
}