Add a range slider to a website

Adds an input to select within the entered range.

Usage

input_slider(label, id = label, ..., min = 0, max = 1, step = 1,
  default = max, note = NULL, dataset = NULL, depends = NULL)

Arguments

label

Label of the input for the user.

id

Unique id of the element to be created.

...

Additional attributes to set on the element.

min

The smallest value in the range.

max

The largest value in the range.

step

How much moving the handle adjusts the selected value.

default

Starting value of the slider handle.

note

Text to display as a tooltip for the input.

dataset

The name of an included dataset, where variable should be looked for; only applies when there are multiple datasets with the same variable name.

depends

The id of another input on which the options depend; this will take president over dataset and variable, depending on this type of input depends points to.

Value

A character vector of the contents to be added.

Examples

if (FALSE) { # \dontrun{
input_slider()
} # }