Add a plot to a webpage

Adds a Plotly plot to a webpage, based on specified or selected .

output_plot(x = NULL, y = NULL, color = NULL, color_time = NULL,
  dataview = NULL, click = NULL, subto = NULL, options = NULL)

Arguments

x
Default title at the top of the element.

y
Default message below the title, to be replaced by the information table.

color
The name of a variable, or ID of a variable selector to use to color lines.

color_time
The ID of a selector to specify which timepoint of color to use.

dataview
The ID of an input_dataview component.

click
The ID of an input to set to a clicked line's ID.

subto
A vector of output IDs to receive hover events from.

options
A list of configuration options, with named entries for any of data, layout, or options, potentially extracted from a saved plotly object (see Plotly documentation).

Value

A character vector of the content to be added.

Examples

if (FALSE) {
output_info("Initial State", "Make a selection for more information.")
}