Adds a menu to a website

Adds a section to arrange inputs, that can optionally be collapsed, and positioned on any side.

Usage

page_menu(..., position = "right", width = "300px", height = NULL,
  collapsible = TRUE, default_open = FALSE, wraps = TRUE, sizes = NA,
  breakpoints = NA, conditions = "")

Arguments

...

Elements to appear in the menu.

position

A character indicating which side to place the menu on; one of "top", "right", "bottom", or "left".

width

Width of the menu, if position is "left" or "right".

height

Height of the menu, if position is "top" or "bottom".

collapsible

Logical; if FALSE, removes the button to open or close the menu.

default_open

Logical; if TRUE, the menu will start open.

wraps

Logical or character ("row" or "col"); if FALSE, elements of the menu will not be wrapped in row or column elements.

sizes

The relative size of each wrapper, between 1 and 12, or "auto"; default is equal size.

breakpoints

Bootstrap breakpoint of each wrapper; one of "" (extra small), "sm",

conditions

A character for each element representing the conditions in which that should be showing (e.g., c("", "input_a == a", "")); "" means the element's display is not conditional. Adding "lock: " before the condition will disable inputs rather than hide the element.

Value

A character vector of the content to be added.

Examples

if (FALSE) { # \dontrun{
page_menu(
  input_select("menu item", c("a", "b", "c"))
)
} # }