Adds an organizational panel to a website

Adds a panel to a website outside of the main content area.

page_panel(title = "Side Panel", ..., position = "left", wraps = NA,
  sizes = NA, breakpoints = NA, conditions = "", id = NULL)

Arguments

title

Content to appear in the panel's header area.

...

Elements to appear in the panel's body area.

position

The side of the screen on which the panel appears; "left" (default) or "right".

wraps

The class of wrapper to place elements in; either "row", "col", or "" (to not wrap the element). Can specify 1 for every element, or a different class for each element.

sizes

The relative size of each wrapper, between 1 and 12, or "auto"; default is equal size.

breakpoints

Bootstrap breakpoint of each wrapper; one of "" (extra small), "sm", "md", "lg", "xl", or "xxl".

conditions

A character for each element representing the conditions in which that should be showing (e.g., c("", "input_a == a", "")); "" means the element's display is not conditional.

id

Unique ID of the section.

Value

A character vector of the content to be added.

Details

See the Bootstrap grid documentation.

Examples

if (FALSE) {
page_panel(
  "<h1>Title</h1>",
  "<p>body</p>",
)
}