Adds an organizational section to a website

Adds a section (such as a row or column) to a website, optionally placing its elements within rows or columns.

Usage

page_section(..., type = "row", wraps = NA, sizes = NA,
  breakpoints = NA, conditions = "", id = NULL)

Arguments

...

Elements to appear in the section.

type

The class of the top-level section; usually either "row" or "col"; NULL for no class.

wraps

The class of wrapper to place elements in; either "row", "col", or "" (to not wrap the element). Can specify 1 for every element, or a different class for each element.

sizes

The relative size of each wrapper, between 1 and 12, or "auto"; default is equal size.

breakpoints

Bootstrap breakpoint of each wrapper; one of "" (extra small), "sm", "md", "lg", "xl", or "xxl".

conditions

A character for each element representing the conditions in which that should be showing (e.g., c("", "input_a == a", "")); "" means the element's display is not conditional. Adding "lock: " before the condition will disable inputs rather than hide the element.

id

Unique ID of the section.

Value

A character vector of the content to be added.

Details

See the Bootstrap grid documentation.

Examples

if (FALSE) { # \dontrun{
page_section(
  "<p>column</p>",
  "<p>row</p>",
  type = "row",
  wraps = c("col", "row")
)
} # }