Reformat an SDAD-formatted dataset

Unify multiple files, which each contain a tall set of variables associated with regions.

data_reformat_sdad(files, value = "value", value_name = "measure",
  id = "geoid", time = "year", dataset = "region_type",
  value_info = "measure_type", entity_info = c("region_type",
  "region_name"), out = NULL)

Arguments

files
A character vector of file paths, or the path to a directory containing data files.

value
Name of the column containing variable values.

value_name
Name of the column containing variable names; assumed to be a single variable per file if not present.

id
Column name IDs which uniquely identify entities.

time
Column name of the variable representing time.

dataset
Column name used to separate entity scales.

value_info
A vector of column names which provide additional information about values (such as their name and type).

entity_info
A vector of column names which provide additional information about entities (such as their name and type).

out
Path to a directory to write files to; if not specified, files will not be written.

Value

An invisible list of the unified variable datasets, split into datasets.

Details

The basic assumption is that there are (a) entities which (b) exist in a hierarchy, and (c1) have a static set of features and (c2) a set of variable features which (d) are assessed at multiple time points.

For example (and generally), entities are (a) regions, with (b) smaller regions making up larger regions, and which (c1) have names, and (c2) population and demographic counts (d) between 2009 and 2019.

Examples

if (FALSE) {
data_reformat_sdad("data/directory")
}