Reformat an SDAD-formatted dataset

Unify multiple files, which each contain a tall set of variables associated with regions.

data_reformat_sdad(files, out = NULL, variables = NULL, ids = NULL,
  value = "value", value_name = "measure", id = "geoid", time = "year",
  dataset = "region_type", value_info = "measure_type",
  entity_info = c(type = "region_type", name = "region_name"),
  formatters = NULL, compression = "xz", read_existing = TRUE,
  overwrite = FALSE, get_coverage = TRUE, verbose = TRUE)

Arguments

files

A character vector of file paths, or the path to a directory containing data files.

out

Path to a directory to write files to; if not specified, files will not be written.

variables

Vector of variable names (in the value_name column) to be included.

ids

Vector of IDs (in the id column) to be included.

value

Name of the column containing variable values.

value_name

Name of the column containing variable names; assumed to be a single variable per file if not present.

id

Column name IDs which uniquely identify entities.

time

Column name of the variable representing time.

dataset

Column name used to separate entity scales.

value_info

A vector of column names which provide additional information about values (such as their name and type).

entity_info

A list containing variable names to extract and create an ids map from ( entity_info.json, created in the output directory). Entries can be named to rename the variables they refer to in entity features.

formatters

A list of functions to pass columns through, with names identifying those columns (e.g., list(region_name = function(x) sub(",.*$", "", x)) to strip text after a comma in the "region_name" column).

compression

A character specifying the type of compression to use on the created files, between "gzip", "bzip2", and "xz". Set to FALSE to disable compression.

read_existing

Logical; if FALSE, will not read in existing sets.

overwrite

Logical; if TRUE, will overwrite existing reformatted files, even if the source files are older than it.

get_coverage

Logical; if FALSE, will not calculate a summary of variable coverage (coverage.csv).

verbose

Logical; if FALSE, will not print status messages.

Value

An invisible list of the unified variable datasets, split into datasets.

Details

The basic assumption is that there are (a) entities which (b) exist in a hierarchy, and (c1) have a static set of features and (c2) a set of variable features which (d) are assessed at multiple time points.

For example (and generally), entities are (a) regions, with (b) smaller regions making up larger regions, and which (c1) have names, and (c2) population and demographic counts (d) between 2009 and 2019.

Examples

if (FALSE) {
data_reformat_sdad("data/directory")
}