Add a button to a website

Adds a button to reset an input or trigger a predefined function to a website.

input_button(label, target = "reset_selection", id = label, class = "", ...)

Arguments

label
Text in the button for the user.

target
The ID of an input element to reset, or the name of a function to trigger. Available function are refresh (to force a reprocessing of the current input state), reset_selection (to reset all input elements that are not settings), and reset_storage (to clear local storage, reset all settings, and reload).

id
Unique ID of the element to be created.

class
Additional class names to add to the element.

...
Additional attributes to set on the element.

Value

A character vector of the contents to be added.

See also

Examples

if (FALSE) {
input_button("Reset")
}