Add a select input to a website

Adds an input to select from the entered options.

input_select(label, options, default = -1, display = options, id = label,
  variable = NULL, dataset = NULL, depends = NULL,
  reset_button = FALSE)

Arguments

label
Label of the input for the user.

options
A vector of options, the name of a variable from which to pull levels, or either "datasets" or "variables" to select names of datasets or variables.

default
Which of the options to default to; either its index or value.

display
A display version of the options.

id
Unique id of the element to be created.

variable
The name of a variable from which to get levels (overwritten by depends).

dataset
The name of an included dataset, where variable should be looked for; only applies when there are multiple datasets with the same variable name.

depends
The id of another input on which the options depend; this will take president over dataset and variable, depending on this type of input depends points to.

reset_button
If specified, adds a button after the select element that will revert the selection to its default; either TRUE, or text for the reset button's label.

Value

A character vector of the contents to be added.

Examples

if (FALSE) {
input_select()
}