Adds a menu to a website

Adds a section to arrange inputs, that can optionally be collapsed, and positioned on any side.

page_menu(..., position = "right", width = NULL, height = NULL,
  collapsible = TRUE, default_open = FALSE, wraps = TRUE, sizes = NA,
  breakpoints = NA)

Arguments

...
Elements to appear in the menu.

position
A character indicating which side to place the menu on; one of "top", "right", "bottom", or "left".

width
Width of the menu, if position if "left" or "right".

height
Height of the menu, if position if "top" or "bottom".

collapsible
Logical; if FALSE, removes the button to open or close the menu.

default_open
Logical; if TRUE, the menu will start open.

wraps
Logical or character ("row" or "col"); if FALSE, elements of the menu will not be wrapped in row or column elements.

sizes
The relative size of each wrapper, between 1 and 12, or "auto"; default is equal size.

breakpoints
Bootstrap breakpoint of each wrapper; one of "" (extra small), "sm",

Value

A character vector of the content to be added.

Examples

if (FALSE) {
page_menu(
  input_select("menu item", c("a", "b", "c"))
)
}